+++ NEWS.md
@@ -1,5 +1,10 @@
 bug reports https://github.com/PredictiveEcology/LandR/issues
 
+# 0.0.2
+
+* new function `overlayLCCs` which will help with overlaying more than one land cover classification raster.
+* major revisions to `convertUnwantedLCC` to accommodate more cases and eliminate redundant arguments.
+
 ### Feb 2nd 2019
 # prepInputsLCC now works with LCC10
 # Bug fixes in imports (DESCRIPTION)